header.css, single-movie.css: Files are refactored to have small class names.
single-rt-movie.php: It will call the header footer and the required template file for the single-rt-movie post type

single-rt-movie-template.php: It is template file that will be called by single-rt-movie.php

cast-crew-template.php: It will have the cast & crew template.

plot-quick-links-template.php: It will have the plot-quick-links part of the website.

poster-title-template.php: It will have the movie poster and title part of the website.